requirements.txt
pandas
psycopg2-binary
requests
pyarrow
jupyter
minio
polars